requirements.txt
easyocr==1.7.1
pytesseract==0.3.10
pillow==10.4.0
